fix restructure error
